chore: move skins folder outside of specific deployment folder
